+++ lib/kb_uploadmethods/kb_uploadmethodsClient.py
@@ -429,6 +429,29 @@ class kb_uploadmethods(object):
             'kb_uploadmethods.import_reads_from_staging',
             [params], self._service_ver, context)
 
+    def import_tsv_as_phenotype_set_from_staging(self, params, context=None):
+        """
+        :param params: instance of type "FileToPhenotypeSetParams" (required
+           params: staging_file_subdir_path: subdirectory file path e.g. for
+           file: /data/bulk/user_name/file_name staging_file_subdir_path is
+           file_name for file:
+           /data/bulk/user_name/subdir_1/subdir_2/file_name
+           staging_file_subdir_path is subdir_1/subdir_2/file_name
+           phenotype_set_name: output PhenotypeSet object name
+           workspace_name: workspace name/ID of the object optional: genome:
+           Genome object that contains features referenced by the Phenotype
+           Set) -> structure: parameter "staging_file_subdir_path" of String,
+           parameter "workspace_name" of type "workspace_name" (workspace
+           name of the object), parameter "phenotype_set_name" of String,
+           parameter "genome" of type "obj_ref"
+        :returns: instance of type "UploadMethodResult" -> structure:
+           parameter "obj_ref" of type "obj_ref", parameter "report_name" of
+           type "report_name", parameter "report_ref" of type "report_ref"
+        """
+        return self._client.call_method(
+            'kb_uploadmethods.import_tsv_as_phenotype_set_from_staging',
+            [params], self._service_ver, context)
+
     def status(self, context=None):
         return self._client.call_method('kb_uploadmethods.status',
                                         [], self._service_ver, context)