+++ lib/src/us/kbase/kbuploadmethods/KbUploadmethodsClient.java
@@ -402,6 +402,23 @@ public class KbUploadmethodsClient {
         return res.get(0);
     }
 
+    /**
+     * <p>Original spec-file function name: import_tsv_as_phenotype_set_from_staging</p>
+     * <pre>
+     * </pre>
+     * @param   params   instance of type {@link us.kbase.kbuploadmethods.FileToPhenotypeSetParams FileToPhenotypeSetParams}
+     * @return   parameter "returnVal" of type {@link us.kbase.kbuploadmethods.UploadMethodResult UploadMethodResult}
+     * @throws IOException if an IO exception occurs
+     * @throws JsonClientException if a JSON RPC exception occurs
+     */
+    public UploadMethodResult importTsvAsPhenotypeSetFromStaging(FileToPhenotypeSetParams params, RpcContext... jsonRpcContext) throws IOException, JsonClientException {
+        List<Object> args = new ArrayList<Object>();
+        args.add(params);
+        TypeReference<List<UploadMethodResult>> retType = new TypeReference<List<UploadMethodResult>>() {};
+        List<UploadMethodResult> res = caller.jsonrpcCall("kb_uploadmethods.import_tsv_as_phenotype_set_from_staging", args, retType, true, true, jsonRpcContext, this.serviceVersion);
+        return res.get(0);
+    }
+
     public Map<String, Object> status(RpcContext... jsonRpcContext) throws IOException, JsonClientException {
         List<Object> args = new ArrayList<Object>();
         TypeReference<List<Map<String, Object>>> retType = new TypeReference<List<Map<String, Object>>>() {};