+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 **1.0.5**:
 Add FBAModel import functions (tsv/excel/sbml)
+added KBasePhenotypes.PhenotypeSet importer
 
 **1.0.3**:
 replace FBAFileUtil with fba_tools module